lcs-sacoche : update to version 2011-06-01c
git-svn-id: https://svn.tice.ac-caen.fr/svn/LcsSE3/LCS/LCS2.0ndg/sources/branches/2.4.8@3974 1012e355-7652-49d3-8fd6-1f9b0b464cea

+++ lcs-sacoche/Sacoche/_sql/structure/sacoche_jointure_user_pilier.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS sacoche_jointure_user_pilier;
 
 CREATE TABLE sacoche_jointure_user_pilier (
 	user_id                MEDIUMINT(8) UNSIGNED                NOT NULL DEFAULT 0,
-	pilier_id              SMALLINT(5)  UNSIGNED                NOT NULL DEFAULT 0,
+	pilier_id              TINYINT(3)   UNSIGNED                NOT NULL DEFAULT 0,
 	validation_pilier_etat TINYINT(1)   UNSIGNED                NOT NULL DEFAULT 1  COMMENT "1 si validation positive ; 0 si validation négative.",
 	validation_pilier_date DATE                                 NOT NULL DEFAULT "0000-00-00",
 	validation_pilier_info VARCHAR(25)  COLLATE utf8_unicode_ci NOT NULL DEFAULT "" COMMENT "Enregistrement statique du nom du validateur, conservé les années suivantes.",
